pyproject.toml
[project]
name = "musciclaim"
version = "0.1.0"
description = "Reproducible evaluation and diagnostics for MuSciClaims."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "MuSciClaim contributors" }]

dependencies = [
  "numpy>=1.26",
  "PyYAML>=6.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
# Hugging Face + multimodal inference stack.
hf = [
  "datasets>=2.20",
  "huggingface_hub>=0.23",
  "Pillow>=10.0",
  "transformers>=4.42",
  "torch",
]

# Dev tools.
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[project.scripts]
musciclaim-check-readmes = "musciclaim.cli.check_readmes:main"
musciclaim-check-docstrings = "musciclaim.cli.check_docstrings:main"
musciclaim-eval = "musciclaim.cli.eval:main"
musciclaim-metrics = "musciclaim.cli.metrics:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
